USEFUL PROCEDURE WITHIN!!! starting with user config for interface
The _to_integer procedure can be called to convert a user input already obtained from a buffer into a hexadecimal int as long as you do
      lea bp, word (buffer here)
      mov bx, (length of string)

I will soon have to move into the loop that goes thru the turn actions and asks the user for a bunch of decisions. I just need to finish the rest of user config.